% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{allparams}
\alias{allparams}
\title{A dump of the final IV and passive cell properties dataframe
This is initally created by CollectCellProperties}
\description{
A dump of the final IV and passive cell properties dataframe
This is initally created by CollectCellProperties
}

