% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{classATAGDB}
\alias{classATAGDB}
\title{Database dump of correspondences between anatomical and functional cell types}
\value{
This data.frame includes the following two important fields: 
  \itemize{
  
  \item \code{class} The cell class defined by its odour response 
  profile/driver line
  
  \item \code{AT} The obsolete Anatomy type (a mid-level classifcation of 
  cells by identify)
  
  \item \code{anatomy.group} The mid-level definition of anatomical type 
  within the final hierarchy of i) primary neurite tract ii) anatomy group, 
  iii) cell type. The anatomy.group is defined by the primary neurite tract 
  and the main tracts taken by the axon and primary dendrite to reach their 
  arborisation zones.
  
  }
}
\description{
This is initally created by DumpDatabase.R
}
