% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{IVfinalDF}
\alias{IVfinalDF}
\title{A dump of the final IV cell properties dataframe
This is initally created by CollectCellProperties}
\format{An object of class \code{data.frame} with 447 rows and 2 columns.}
\usage{
IVfinalDF
}
\description{
A dump of the final IV cell properties dataframe
This is initally created by CollectCellProperties
}
\keyword{datasets}
