% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{IVInpResbig}
\alias{IVInpResbig}
\title{A dump of the final IV cell properties dataframe 3 points/cell
This is initally created by CollectCellProperties}
\description{
A dump of the final IV cell properties dataframe 3 points/cell
This is initally created by CollectCellProperties
}
