% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata-package.R
\docType{package}
\name{physplitdata-package}
\alias{physplitdata}
\alias{physplitdata-package}
\title{Packaged and versioned spikes and metadata for Shahar's cells}
\description{
The main objects are \code{\link{Spikes}}, \code{\link{smSpikes}} and
\code{\link{PhySplitDB}}
}
\details{
Data objects initally created by external code \code{make_physplitdata.R}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferislab/physplitdata}
  \item Report bugs at \url{https://github.com/jefferislab/physplitdata/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

Authors:
\itemize{
  \item Shahar Frechter (0000-0002-0431-5849)
}

}
