% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{package}
\name{physplitdata}
\alias{physplitdata}
\alias{physplitdata-package}
\title{Packaged and versioned spikes and metadata for Shahar's cells}
\description{
The main objects are \code{\link{Spikes}}, \code{\link{smSpikes}} and \code{\link{PhySplitDB}}
}
\details{
This is initally created by make_physplitdata.R
}

