% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{PhySplitDB}
\alias{PhySplitDB}
\title{A dump of key cell metadata from the PhySplit Filemaker database}
\format{An object of class \code{data.frame} with 587 rows and 28 columns.}
\usage{
PhySplitDB
}
\description{
This is initally created by DumpDatabase.R
}
\keyword{datasets}
