% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{ravc}
\alias{ravc}
\title{A dump of passive cell properties}
\format{An object of class \code{list} of length 378.}
\usage{
ravc
}
\description{
This is initally created by CollectCellProperties
}
\keyword{datasets}
