% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{smSpikes}
\alias{smSpikes}
\title{Smoothed PSTH for all spikes from Shahar's cells}
\format{An object of class \code{list} of length 469.}
\usage{
smSpikes
}
\description{
This is initally created by CollectSpikes.R
}
\details{
The psth is calculated for a window 0-3 seconds using a bandwidth of 0.5s and
a time step of 0.05
}
\seealso{
\code{\link{Spikes}}, \code{\link[gphys]{smpsth}},
  \code{\link[STAR]{psth}}
}
\keyword{datasets}
