% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{Vm}
\alias{Vm}
\title{Smoothed average membrane potential for Shahar's cells}
\format{An object of class \code{list} of length 492.}
\usage{
Vm
}
\description{
This is initally created by make_vm.R. Each individual object in the list was
created by the smooth_decimate function.
}
\examples{
\dontrun{
library(gphys)
plot(Spikes[['nm20120605c0']][[1]])
AddLinesToRasterPlot(Vm[['nm20120605c0']], ylim=c(-65,-35), col='darkgreen')
}
}
\seealso{
\code{\link{Spikes}}, \code{\link[gphys]{AddLinesToRasterPlot}}
}
\keyword{datasets}
