% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{allparams}
\alias{allparams}
\title{A dump of the final IV and passive cell properties dataframe
This is initally created by CollectCellProperties}
\format{An object of class \code{data.frame} with 454 rows and 5 columns.}
\usage{
allparams
}
\description{
A dump of the final IV and passive cell properties dataframe
This is initally created by CollectCellProperties
}
\keyword{datasets}
