% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{cpdf.summ}
\alias{cpdf.summ}
\title{A dump of the final passive cell properties dataframe
This is initally created by CollectCellProperties}
\description{
A dump of the final passive cell properties dataframe
This is initally created by CollectCellProperties
}

