% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{IVfinalDF}
\alias{IVfinalDF}
\title{A dump of the final IV cell properties dataframe
This is initally created by CollectCellProperties}
\description{
A dump of the final IV cell properties dataframe
This is initally created by CollectCellProperties
}
