% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{primary_neurites.is2}
\alias{primary_neurites.is2}
\alias{primary_neurites}
\alias{primary_neurites.jfrc2}
\title{Primary neurite tracts (and accompanying metadata)}
\description{
\code{primary_neurites.is2} is in the IS2 template brain space
  (Jefferis Lab).

\code{primary_neurites.jfrc2} is in the JFRC2 template brain
  space (Janelia / Virtual Fly Brain).
}
\details{
This is initally created by \code{data-raw/pct.R} which in turn
  depends on \code{LHNTracing/R/MakeParentClusters.R}.
}
\examples{
head(primary_neurites.is2)
\dontrun{
library(nat.flybrains)
plot3d(primary_neurites.is2, col=factor(finalcluster))
plot3d(IS2)
}
}
\seealso{
\code{\link{lhnsxfinal}}
}
