% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplitdata.R
\docType{data}
\name{Vm}
\alias{Vm}
\title{Smoothed average membrane potential for Shahar's cells}
\description{
This is initally created by make_vm.R. Each individual object in the list was
created by the smooth_decimate function.
}
\examples{
plot(Spikes[[1]][[1]])
AddLinesToRasterPlot(Vm[[1]], ylim=c(-65,-35), col='darkgreen')
}
\seealso{
\code{\link{Spikes}}
}

